# Add benchmark of the Variables::extend() method 