--- a/micro/CMakeLists.txt
+++ b/micro/CMakeLists.txt
@@ -5,6 +5,7 @@ IF (benchmark_FOUND)
 		benchmark.cc
 		scopelink_bm.cc
 		evaluationlink_bm.cc
+		variables_bm.cc
 	)
 
 	TARGET_LINK_LIBRARIES (benchmark