--- a/micro/variables_bm.cc
+++ b/micro/variables_bm.cc
@@ -24,25 +24,81 @@
 
 #include <opencog/util/Logger.h>
 #include <opencog/atomspace/AtomSpace.h>
+#include <opencog/atomspace/AtomSpace.h>
+#include <opencog/atoms/core/Variables.h>
+#include <opencog/atoms/core/VariableList.h>
 
 #include "benchmark.h"
 
 using namespace opencog;
 
-static void BM_ExtendVariablesNewVariable(benchmark::State& state)
+static void BM_VariablesExtend_NewVariable(benchmark::State& state)
 {
 	AtomSpace atomspace;
 	Handle varX = atomspace.add_node(VARIABLE_NODE, "$X");
 	Handle varY = atomspace.add_node(VARIABLE_NODE, "$Y");
-	Handle varDeclA = atomspace.add_link(VARIABLE_LIST, VarX);
-	Handle varDeclB = atomspace.add_link(VARIABLE_LIST, VarY);
-	VariableList vl1(varDeclA), vl2(varDeclB);
+	Handle varDeclA = atomspace.add_link(VARIABLE_LIST, varX);
+	Handle varDeclB = atomspace.add_link(VARIABLE_LIST, varY);
+	Variables varsA(VariableList(varDeclA).get_variables());
+	Variables varsB(VariableList(varDeclB).get_variables());
 
 	for (auto _ : state)
 	{
-		vl1.extend(vl2);
+		varsA.extend(varsB);
 	}
 }
-BENCHMARK(BM_ExtendVariablesNewVariable);
+BENCHMARK(BM_VariablesExtend_NewVariable);
+
+static void BM_VariablesExtend_SameVariableNoTypeRestrictions(benchmark::State& state)
+{
+	AtomSpace atomspace;
+	Handle varX = atomspace.add_node(VARIABLE_NODE, "$X");
+	Handle conceptNodeType = atomspace.add_node(TYPE_NODE, "ConceptNode");
+	Handle varXTyped = atomspace.add_link(TYPED_VARIABLE_LINK, varX, conceptNodeType);
+	Handle varDeclA = atomspace.add_link(VARIABLE_LIST, varXTyped);
+	Handle varDeclB = atomspace.add_link(VARIABLE_LIST, varX);
+	Variables varsA(VariableList(varDeclA).get_variables());
+	Variables varsB(VariableList(varDeclB).get_variables());
+
+	for (auto _ : state)
+	{
+		varsA.extend(varsB);
+	}
+}
+BENCHMARK(BM_VariablesExtend_SameVariableNoTypeRestrictions);
+
+static void BM_VariablesExtend_NewVariableNoTypeRestrictions(benchmark::State& state)
+{
+	AtomSpace atomspace;
+	Handle varX = atomspace.add_node(VARIABLE_NODE, "$X");
+	Handle varDeclA = atomspace.add_link(VARIABLE_LIST);
+	Handle varDeclB = atomspace.add_link(VARIABLE_LIST, varX);
+	Variables varsA(VariableList(varDeclA).get_variables());
+	Variables varsB(VariableList(varDeclB).get_variables());
+
+	for (auto _ : state)
+	{
+		varsA.extend(varsB);
+	}
+}
+BENCHMARK(BM_VariablesExtend_NewVariableNoTypeRestrictions);
+
+static void BM_VariablesExtend_SameVariableWithTypeRestrictions(benchmark::State& state)
+{
+	AtomSpace atomspace;
+	Handle varX = atomspace.add_node(VARIABLE_NODE, "$X");
+	Handle conceptNodeType = atomspace.add_node(TYPE_NODE, "ConceptNode");
+	Handle varXTyped = atomspace.add_link(TYPED_VARIABLE_LINK, varX, conceptNodeType);
+	Handle varDeclA = atomspace.add_link(VARIABLE_LIST, varXTyped);
+	Handle varDeclB = atomspace.add_link(VARIABLE_LIST, varXTyped);
+	Variables varsA(VariableList(varDeclA).get_variables());
+	Variables varsB(VariableList(varDeclB).get_variables());
+
+	for (auto _ : state)
+	{
+		varsA.extend(varsB);
+	}
+}
+BENCHMARK(BM_VariablesExtend_SameVariableWithTypeRestrictions);