Benchmark for the https://github.com/opencog/atomspace/pull/2427

Results before change:
```
$ ./micro/benchmark --benchmark_filter=BM_VariablesExtend
2019-12-10 18:08:19
Running ./micro/benchmark
Run on (12 X 4000 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 0.77, 0.61, 0.65
----------------------------------------------------------------------------------------------
Benchmark                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------
BM_VariablesExtend_NewVariable                            1880 ns         1880 ns       372825
BM_VariablesExtend_SameVariableNoTypeRestrictions         1871 ns         1871 ns       371455
BM_VariablesExtend_NewVariableNoTypeRestrictions          1864 ns         1864 ns       374311
BM_VariablesExtend_SameVariableWithTypeRestrictions        140 ns          140 ns      4958600
```

After change:
```
$ ./micro/benchmark --benchmark_filter=BM_VariablesExtend
2019-12-10 18:09:12
Running ./micro/benchmark
Run on (12 X 4000 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 1.08, 0.73, 0.69
----------------------------------------------------------------------------------------------
Benchmark                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------
BM_VariablesExtend_NewVariable                            14.7 ns         14.7 ns     47369143
BM_VariablesExtend_SameVariableNoTypeRestrictions         12.0 ns         12.0 ns     58611996
BM_VariablesExtend_NewVariableNoTypeRestrictions          10.8 ns         10.8 ns     65308978
BM_VariablesExtend_SameVariableWithTypeRestrictions        146 ns          146 ns      4800956
```